requirements.txt
streamlit
torch
torchvision
ultralytics
pillow
opencv-python-headless
